refactor: rename dialog/style.ts to dialog/styles/ts
