+++ src/components/dialog/styles.ts
